requirements.txt
# Job Finder - Python Dependencies
# Install into the project venv:
#   pip install -r requirements.txt

# ---------------------------------------------------------------------------
# Core already installed in the project venv (listed for reference)
# ---------------------------------------------------------------------------

# PDF parsing
pypdf>=4.0.0

# HTTP requests and HTML parsing
requests>=2.31.0
beautifulsoup4>=4.12.0

# Data manipulation
pandas>=2.0.0
numpy>=1.24.0

# Google Sheets / Drive API
google-api-python-client>=2.100.0
google-auth>=2.23.0
google-auth-httplib2>=0.1.1
google-auth-oauthlib>=1.1.0
googleapis-common-protos>=1.60.0

# ---------------------------------------------------------------------------
# Scheduler
# ---------------------------------------------------------------------------

APScheduler>=3.10.0

# ---------------------------------------------------------------------------
# NLP / matching (lightweight – no sklearn required; brain.py is pure Python)
# Optional: uncomment if you want scikit-learn based TF-IDF
# ---------------------------------------------------------------------------

# scikit-learn>=1.3.0    # optional: for sklearn TF-IDF / cosine_similarity
# fuzzywuzzy>=0.18.0     # optional: fuzzy string matching
# python-Levenshtein>=0.21.0  # speeds up fuzzywuzzy

# ---------------------------------------------------------------------------
# Environment / configuration
# ---------------------------------------------------------------------------

python-dotenv>=1.0.0

# ---------------------------------------------------------------------------
# Optional: browser automation for anti-bot job boards (production mode)
# ---------------------------------------------------------------------------

# selenium>=4.15.0          # browser automation (LinkedIn / Indeed production)
# webdriver-manager>=4.0.0  # auto-downloads chromedriver

# ---------------------------------------------------------------------------
# Optional: Twilio SMS notifications (already installed in venv)
# ---------------------------------------------------------------------------

# twilio>=8.0.0   # already present in venv

# ---------------------------------------------------------------------------
# Development / testing
# ---------------------------------------------------------------------------

# pytest>=7.4.0
# pytest-mock>=3.11.0
